fix(refactor): dropped parsers file, drop handle both drop and pop
