+++ src/args.rs
@@ -4,7 +4,7 @@
 //! It will ensure we get the correct args and then return
 //! a correct Structure to run the corresponding commands
 use crate::{
-    commands::{RunnableCommand, drop, list, pop, record, run},
+    commands::{RunnableCommand, drop, list, record, run},
     errors::ReplayError,
 };
 use clap::{Parser, Subcommand};
@@ -27,10 +27,7 @@ pub enum CliCommand {
     /// List all the sessions recorded
     List(list::ListCommand),
 
-    /// Remove the laste session recorded
-    Pop(pop::PopCommand),
-
-    /// Drop a specified session
+    /// Drop a specified session, last session if not specified
     Drop(drop::DropCommand),
 }
 
@@ -40,7 +37,6 @@ impl CliCommand {
             CliCommand::Run(cmd) => cmd.run(),
             CliCommand::Record(cmd) => cmd.run(),
             CliCommand::List(cmd) => cmd.run(),
-            CliCommand::Pop(cmd) => cmd.run(),
             CliCommand::Drop(cmd) => cmd.run(),
         }
     }
@@ -49,6 +45,19 @@ impl CliCommand {
 pub fn parse_command(args: &[String]) -> Result<CliCommand, ReplayError> {
     let cli_command = CliParser::try_parse_from(args)?;
     Ok(cli_command.command)
+}
+
+pub fn parse_session_index(s: &str) -> Result<u32, String> {
+    s.strip_prefix("replay@{")
+        .and_then(|rest| rest.strip_suffix('}'))
+        .ok_or_else(|| {
+            format!(
+                "Session name must be of the form replay@{{index}}, got '{}'",
+                s
+            )
+        })?
+        .parse::<u32>()
+        .map_err(|_| format!("Invalid session index in '{}'", s))
 }
 
 #[cfg(test)]
